pyproject.toml
[project]
name = "chat-with-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.12",
    "docx2txt>=0.9",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.25",
    "langchain-google-genai>=2.1.6",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.27",
    "pillow>=11.3.0",
    "pypdf>=5.6.0",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
